autofocus, state mutex
Add autofocus option to config. When set to true, new pane is
auto-focused on start. When a pid terminates focus is switched back to
proctmux.

This commit includes significant changes to controller. State is now
wrapped in a mutex. Changes to state are limited to a single method,
lock_and_load. This method locks the mutex and calls a function that may
return a modified state. If a modified state is returned the mutex value
is updated and the screen is redrawn.